move header into separate component
