+++ src/content/app/tools/blast/components/blast-submission-header-container/BlastSubmissionHeaderGrid.scss
