+++ src/content/app/tools/blast/components/listed-blast-submission/ListedBlastSubmission.tsx
@@ -129,7 +129,7 @@ const CollapsedSequencesBox = (props: Props) => {
   );
 };
 
-export const Header = (
+const Header = (
   props: Props & {
     isAnyJobRunning: boolean;
     isExpanded: boolean;
